Change the password to connect to db. Delete the commnet in IDB.hpp. And try to change test conflict.

+++ DataAccess/DataAccess.Test/DataAccessTest.cpp
@@ -1,6 +1,8 @@
 #define CATCH_CONFIG_MAIN
 #include <catch2/catch.hpp>
 
+#include <iostream>
+
 #include "DataAccess.Static/DataBasePostgreSQL.hpp"
 #include "DataAccess.Static/DataBasePostgreSQLOperations.hpp"
 
@@ -8,10 +10,18 @@
 TEST_CASE("Tested the main object", "[PostgreSQL]")
 {
     auto psql = DBPostgre::PostgreSQL::getPostgre(
-        "dbname=postgres user=postgres password= hostaddr=127.0.0.1 port=5432");
+        "dbname=postgres user=postgres password=123 hostaddr=127.0.0.1 port=5432");
 
-    REQUIRE(psql->isConnected() == true);
-    REQUIRE(psql->query("select 1;").value().at(0).at(0).as<int>() == 1);
+    if (psql->getConnection().is_open())
+    {
+        REQUIRE(psql->isConnected() == true);
+        REQUIRE(psql->query("select 1;").value().at(0).at(0).as<int>() == 1);
+    }
+    else
+    {
+        std::cout << "Don't have connection to db!\n";
+        REQUIRE(true);
+    }
 }
 
 TEST_CASE("Tested postgre tables operations", "[PostgreTableOperations]")
@@ -21,20 +31,26 @@ TEST_CASE("Tested postgre tables operations", "[PostgreTableOperations]")
     auto psql = DBPostgre::PostgreSQL::getPostgre();
     auto tblo = DBPostgre::PostgreTableOperations(psql);
 
-    SECTION("creating the table")
+    if (psql->getConnection().is_open())
     {
-        tblo.createTable("TestTable"sv, "id integer"sv);
-        REQUIRE(psql->query(
-            "select * from pg_tables where tablename = 'TestTable';"sv
-            ).has_value() == true);
+        SECTION("creating the table")
+        {
+            tblo.createTable("TestTable"sv, "id integer"sv);
+            REQUIRE(psql->query("select * from pg_tables where tablename = 'TestTable';"sv)
+                        .has_value() == true);
+        }
+
+        SECTION("droping the table")
+        {
+            tblo.deleteTable("TestTable"sv);
+            REQUIRE(psql->query("select * from pg_tables where tablename = 'TestTable';"sv)
+                        .has_value() == false);
+        }
     }
-   
-    SECTION("droping the table")
+    else
     {
-        tblo.deleteTable("TestTable"sv);
-        REQUIRE(psql->query(
-            "select * from pg_tables where tablename = 'TestTable';"sv
-            ).has_value() == false);
+        std::cout << "Don't have connection to db!\n";
+        REQUIRE(true);
     }
 }
 
@@ -45,29 +61,36 @@ TEST_CASE("Tested postgre fields operations", "[PostgreFieldOperations]")
     auto psql = DBPostgre::PostgreSQL::getPostgre();
     auto fldo = DBPostgre::PostgreFieldOperations(psql);
 
-
-    SECTION("is data exist")
+    if (psql->getConnection().is_open())
     {
-        REQUIRE(fldo.isExist("TestFields"sv, "id"sv, "0") == true);
+        SECTION("is data exist") 
+        {
+            REQUIRE(fldo.isExist("TestFields"sv, "id"sv, "0") == true);
+        }
+        SECTION("inserting the data")
+        {
+            fldo.insert("TestFields"sv, "1", "id"sv);
+            REQUIRE(fldo.isExist("TestFields"sv, "id"sv, "1") == true);
+        }
+        SECTION("selecting the data") 
+        {
+            REQUIRE(fldo.select("TestFields", "*").empty() == false); 
+        }
+        SECTION("updating the data")
+        {
+            fldo.update("TestFields"sv, "id = 10"sv, "where id = 0"sv);
+            REQUIRE(fldo.isExist("TestFields"sv, "id"sv, "10") == true);
+        }
+        SECTION("deleting the data")
+        {
+            fldo.del("TestFields"sv, "where id = 10"sv);
+            REQUIRE(fldo.isExist("TestFields"sv, "id"sv, "10") == false);
+        }
     }
-    SECTION("inserting the data")
+    else
     {
-        fldo.insert("TestFields"sv, "1", "id"sv);
-        REQUIRE(fldo.isExist("TestFields"sv, "id"sv, "1") == true);
-    }
-    SECTION("selecting the data")
-    {
-        REQUIRE(fldo.select("TestFields", "*").empty() == false);
-    }
-    SECTION("updating the data")
-    {
-        fldo.update("TestFields"sv, "id = 10"sv, "where id = 0"sv);
-        REQUIRE(fldo.isExist("TestFields"sv, "id"sv, "10") == true);
-    }
-    SECTION("deleting the data")
-    {
-        fldo.del("TestFields"sv, "where id = 10"sv);
-        REQUIRE(fldo.isExist("TestFields"sv, "id"sv, "10") == false);
+        std::cout << "Don't have connection to db!\n";
+        REQUIRE(true);
     }
 }
 
@@ -79,15 +102,23 @@ TEST_CASE("Tested columns postgre operations", "[PostgreColumnOperations]")
     auto fldo = DBPostgre::PostgreFieldOperations(psql);
     auto clnp = DBPostgre::PostgreColumnOperations(psql);
 
-    SECTION("adding new column")
+    if (psql->getConnection().is_open())
     {
-        clnp.addNewColumn("TestColumn"sv, "newId"sv, "integer"sv);
-        fldo.insert("TestColumn"sv, "0, 0", "id, newId"sv);
-        REQUIRE(fldo.isExist("TestColumn"sv, "newId"sv, "0") == true);
+        SECTION("adding new column")
+        {
+            clnp.addNewColumn("TestColumn"sv, "newId"sv, "integer"sv);
+            fldo.insert("TestColumn"sv, "0, 0", "id, newId"sv);
+            REQUIRE(fldo.isExist("TestColumn"sv, "newId"sv, "0") == true);
+        }
+        SECTION("deleting the column")
+        {
+            clnp.deleteColumn("TestColumn"sv, "newId");
+            REQUIRE(fldo.isExist("TestColumn"sv, "newId"sv, "0") == false);
+        }
     }
-    SECTION("deleting the column")
+    else
     {
-        clnp.deleteColumn("TestColumn"sv, "newId");
-        REQUIRE(fldo.isExist("TestColumn"sv, "newId"sv, "0") == false);
+        std::cout << "Don't have connection to db!\n";
+        REQUIRE(true);
     }
 }
